OPENTOK-40778: Official OpenTok samples for Linux SDK
