--- Basic-Video-Chat-With-Server/CMakeLists.txt
@@ -1,24 +1,0 @@
-PROJECT(basic_video_chat_with_server)
-
-SET(CMAKE_VERBOSE_MAKEFILE ON)
-
-CMAKE_MINIMUM_REQUIRED(VERSION 3.5)
-
-SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11 -stdlib=libc++")
-
-# Remove the commented out line below to enable OpenTok debug traces.
-# SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DDEBUG")
-
-INCLUDE(FindPkgConfig)
-PKG_SEARCH_MODULE(LIBCURL REQUIRED libcurl)
-PKG_SEARCH_MODULE(LIBJANSSON REQUIRED jansson)
-PKG_SEARCH_MODULE(LIBUNWIND REQUIRED libunwind)
-PKG_SEARCH_MODULE(LIBUV REQUIRED libuv)
-PKG_SEARCH_MODULE(LIBSDL2 REQUIRED sdl2)
-PKG_SEARCH_MODULE(LIBOPENTOK REQUIRED libopentok)
-
-INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR}/../src ${LIBCURL_INCLUDE_DIRS} ${LIBJANSSON_INCLUDE_DIRS} ${LIBUNWIND_INCLUDE_DIRS} ${LIBUV_INCLUDE_DIRS} ${LIBSDL2_INCLUDE_DIRS} ${LIBOPENTOK_INCLUDE_DIRS})
-LINK_DIRECTORIES(${LIBCURL_LIBRARY_DIRS} ${LIBJANSSON_LIBRARY_DIRS} ${LIBUNWIND_LIBRARY_DIRS} ${LIBUV_LIBRARY_DIRS} ${LIBSDL2_LIBRARY_DIRS} ${LIBOPENTOK_LIBRARY_DIRS})
-
-ADD_EXECUTABLE(${PROJECT_NAME} main.cpp ${CMAKE_CURRENT_SOURCE_DIR}/../src/conference.cpp ${CMAKE_CURRENT_SOURCE_DIR}/../src/credential_service_provider_helper.cpp ${CMAKE_CURRENT_SOURCE_DIR}/../src/renderer.cpp)
-TARGET_LINK_LIBRARIES(${PROJECT_NAME} ${LIBCURL_LIBRARIES} ${LIBJANSSON_LIBRARIES} ${LIBUV_LIBRARIES} ${LIBUNWIND_LIBRARIES} ${LIBSDL2_LIBRARIES} ${LIBOPENTOK_LIBRARIES})
