+++ Basic-Video-Chat/main.cpp
@@ -1,47 +1,253 @@
+#include <opentok.h>
+
+#include <algorithm>
 #include <cstdlib>
+#include <iterator>
 #include <iostream>
+#include <map>
 #include <string>
 
-#include "conference.h"
 #include "renderer.h"
 
-#define SIGBREAK 21
+#define API_KEY ""
+#define SESSION_ID ""
+#define TOKEN ""
 
-#define PUBLISHER_NAME "OpenTok Linux SDK Sample - Basic Video Chat"
+static otc_bool g_is_connected = OTC_FALSE;
+static otc_publisher *g_publisher = nullptr;
+static otc_bool g_is_publishing = OTC_FALSE;
+static std::map<std::string, void*> g_subscriber_map;
 
-struct MyConferenceObserver: public ConferenceObserver {
-  void onConnected() {
-    std::cout << "MyConferenceObserver::onConnected()" << std::endl;
+static void on_subscriber_connected(otc_subscriber *subscriber,
+                                    void *user_data,
+                                    const otc_stream *stream) {
+  std::cout << __FUNCTION__ << " callback function" << std::endl;
+}
+
+static void on_subscriber_render_frame(otc_subscriber *subscriber,
+                                       void *user_data,
+                                       const otc_video_frame *frame) {
+  RendererManager *render_manager = static_cast<RendererManager*>(user_data);
+  if (render_manager == nullptr) {
+    return;
+  }
+  render_manager->addFrame(subscriber, frame);
+}
+
+static void on_subscriber_error(otc_subscriber* subscriber,
+                                void *user_data,
+                                const char* error_string,
+                                enum otc_subscriber_error_code error) {
+  std::cout << __FUNCTION__ << " callback function" << std::endl;
+  std::cout << "Subscriber error. Error code: " << error_string << std::endl;
+}
+
+static void on_session_connected(otc_session *session, void *user_data) {
+  std::cout << __FUNCTION__ << " callback function" << std::endl;
+
+  g_is_connected = OTC_TRUE;
+
+  if ((session != nullptr) && (g_publisher != nullptr)) {
+    if (otc_session_publish(session, g_publisher) == OTC_SUCCESS) {
+      g_is_publishing = OTC_TRUE;
+      return;
+    }
+    std::cout << "Could not publish successfully" << std::endl;
+  }
+}
+
+static void on_session_connection_created(otc_session *session,
+                                          void *user_data,
+                                          const otc_connection *connection) {
+  std::cout << __FUNCTION__ << " callback function" << std::endl;
+}
+
+static void on_session_connection_dropped(otc_session *session,
+                                          void *user_data,
+                                          const otc_connection *connection) {
+  std::cout << __FUNCTION__ << " callback function" << std::endl;
+}
+
+static void on_session_stream_received(otc_session *session,
+                                       void *user_data,
+                                       const otc_stream *stream) {
+  std::cout << __FUNCTION__ << " callback function" << std::endl;
+  RendererManager *render_manager = static_cast<RendererManager*>(user_data);
+  if (render_manager == nullptr) {
+    return;
   }
 
-  void onInCall() {
-    std::cout << "MyConferenceObserver::onInCall()" << std::endl;
+  struct otc_subscriber_callbacks subscriber_callbacks = {0};
+  subscriber_callbacks.user_data = user_data;
+  subscriber_callbacks.on_connected = on_subscriber_connected;
+  subscriber_callbacks.on_render_frame = on_subscriber_render_frame;
+  subscriber_callbacks.on_error = on_subscriber_error;
+
+  otc_subscriber *subscriber = otc_subscriber_new(stream,
+                                                  &subscriber_callbacks);
+  if (subscriber == nullptr) {
+    std::cout << "Could not create OpenTok subscriber successfuly" << std::endl;
+    return;
+  }
+  render_manager->createRenderer(subscriber);
+  if (otc_session_subscribe(session, subscriber) == OTC_SUCCESS) {
+    g_subscriber_map[std::string(otc_stream_get_id(stream))] = subscriber;
+    return;
+  }
+  std::cout << "Could not subscribe successfully" << std::endl;
+}
+
+static void on_session_stream_dropped(otc_session *session,
+                                      void *user_data,
+                                      const otc_stream *stream) {
+  std::cout << __FUNCTION__ << " callback function" << std::endl;
+  otc_subscriber *subscriber = static_cast<otc_subscriber *>(g_subscriber_map[std::string(otc_stream_get_id(stream))]);
+  if (subscriber == nullptr) {
+    return;
+  }
+  
+  RendererManager *render_manager = static_cast<RendererManager*>(user_data);
+  if (render_manager != nullptr) {
+    render_manager->destroyRenderer(subscriber);
+  }
+  otc_subscriber_delete(subscriber);
+  g_subscriber_map[std::string(otc_stream_get_id(stream))] = nullptr;
+  subscriber = nullptr;
+}
+
+static void on_session_disconnected(otc_session *session, void *user_data) {
+  std::cout << __FUNCTION__ << " callback function" << std::endl;
+}
+
+static void on_session_error(otc_session *session,
+                             void *user_data,
+                             const char *error_string,
+                             enum otc_session_error_code error) {
+  std::cout << __FUNCTION__ << " callback function" << std::endl;
+  std::cout << "Session error. Error : " << error_string << std::endl;
+}
+
+static void on_publisher_stream_created(otc_publisher *publisher,
+                                        void *user_data,
+                                        const otc_stream *stream) {
+  std::cout << __FUNCTION__ << " callback function" << std::endl;
+}
+
+static void on_publisher_render_frame(otc_publisher *publisher,
+                                      void *user_data,
+                                      const otc_video_frame *frame) {
+  RendererManager *render_manager = static_cast<RendererManager*>(user_data);
+  if (render_manager == nullptr) {
+    return;
+  }
+  render_manager->addFrame(publisher, frame);
+}
+
+static void on_publisher_stream_destroyed(otc_publisher *publisher,
+                                          void *user_data,
+                                          const otc_stream *stream) {
+  std::cout << __FUNCTION__ << " callback function" << std::endl;
+}
+
+static void on_publisher_error(otc_publisher *publisher,
+                               void *user_data,
+                               const char* error_string,
+                               enum otc_publisher_error_code error_code) {
+  std::cout << __FUNCTION__ << " callback function" << std::endl;
+  std::cout << "Publisher error. Error code: " << error_string << std::endl;
+}
+
+static void on_otc_log_message(const char* message) {
+  std::cout <<  __FUNCTION__ << ":" << message << std::endl;
+}
+
+int main(int argc, char** argv) {
+  if (otc_init(nullptr) != OTC_SUCCESS) {
+    std::cout << "Could not init OpenTok library" << std::endl;
+    return EXIT_FAILURE;
+  }
+  otc_log_set_logger_callback(on_otc_log_message);
+  otc_log_enable(OTC_LOG_LEVEL_INFO);
+
+  RendererManager renderer_manager;
+
+  struct otc_session_callbacks session_callbacks = {0};
+  session_callbacks.user_data = &renderer_manager;
+  session_callbacks.on_connected = on_session_connected;
+  session_callbacks.on_connection_created = on_session_connection_created;
+  session_callbacks.on_connection_dropped = on_session_connection_dropped;
+  session_callbacks.on_stream_received = on_session_stream_received;
+  session_callbacks.on_stream_dropped = on_session_stream_dropped;
+  session_callbacks.on_disconnected = on_session_disconnected;
+  session_callbacks.on_error = on_session_error;
+
+  otc_session *session = nullptr;
+  session = otc_session_new(API_KEY, SESSION_ID, &session_callbacks);
+
+  if (session == nullptr) {
+    std::cout << "Could not create OpenTok session successfuly" << std::endl;
+    return EXIT_FAILURE;
   }
 
-  void onNewRemoteParticipant() {
-    std::cout << "MyConferenceObserver::onNewRemoteParticipant()" << std::endl;
+  struct otc_publisher_callbacks publisher_callbacks = {0};
+  publisher_callbacks.user_data = &renderer_manager;
+  publisher_callbacks.on_stream_created = on_publisher_stream_created;
+  publisher_callbacks.on_render_frame = on_publisher_render_frame;
+  publisher_callbacks.on_stream_destroyed = on_publisher_stream_destroyed;
+  publisher_callbacks.on_error = on_publisher_error;
+
+  g_publisher = otc_publisher_new("name",
+                                  nullptr, /* Use WebRTC's video capturer. */
+                                  &publisher_callbacks);
+  if (g_publisher == nullptr) {
+    std::cout << "Could not create OpenTok publisher successfuly" << std::endl;
+    otc_session_delete(session);
+    return EXIT_FAILURE;
+  }
+  renderer_manager.createRenderer(g_publisher);
+
+  otc_session_connect(session, TOKEN);
+
+  renderer_manager.runEventLoop();
+
+  if (session != nullptr) {
+    std::for_each(g_subscriber_map.begin(),
+                  g_subscriber_map.end(),
+                  [&renderer_manager, session](std::pair<std::string, void*> element) {
+      otc_subscriber *subscriber = static_cast<otc_subscriber *>(element.second);
+      if (subscriber == nullptr) {
+        return;
+      }
+      renderer_manager.destroyRenderer(subscriber);
+      otc_session_unsubscribe(session, subscriber);
+      otc_subscriber_delete(subscriber);
+      subscriber = nullptr;
+    });
   }
 
-  void onError(const std::string& errorMessage) {
-    std::cout << "MyConferenceObserver::onError() - " << errorMessage << std::endl;
+  renderer_manager.destroyRenderer(g_publisher);
+
+  if ((session != nullptr) && (g_publisher != nullptr) && (g_is_publishing == OTC_TRUE)) {
+    otc_session_unpublish(session, g_publisher);
+    g_is_publishing = OTC_FALSE;
   }
-};
 
-int main(int argc, char* argv[]) {
-  std::string apiKey = "";
-  std::string sessionId = "";
-  std::string token = "";
+  if (g_publisher != nullptr) {
+    otc_publisher_delete(g_publisher);
+    g_publisher = nullptr;
+  }
 
-  std::shared_ptr<MyConferenceObserver> observer = std::make_shared<MyConferenceObserver>();
-  std::shared_ptr<RendererManager> rendererManager = std::make_shared<RendererManager>();
-  std::unique_ptr<Conference> conference(new Conference(std::string(PUBLISHER_NAME), observer, rendererManager));
-  conference->join(apiKey, sessionId, token);
+  if ((session != nullptr) && (g_is_connected == OTC_TRUE)) {
+    otc_session_disconnect(session);
+    g_is_connected = OTC_FALSE;
+  }
 
-  rendererManager->runEventLoop();
+  if (session != nullptr) {
+    otc_session_delete(session);
+    session = nullptr;
+  }
 
-  conference->leave();
-
-  std::cout << "Quiting..." << std::endl;
+  otc_destroy();
 
   return EXIT_SUCCESS;
 }
