+++ Custom-Video-Capturer/CMakeLists.txt
@@ -1,4 +1,4 @@
-PROJECT(basic_video_chat)
+PROJECT(custom-video-capturer)
 
 SET(CMAKE_VERBOSE_MAKEFILE ON)
 
@@ -18,5 +18,5 @@ FIND_PACKAGE(Threads)
 INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR}/../common/src ${LIBSDL2_INCLUDE_DIRS} ${LIBOPENTOK_INCLUDE_DIRS})
 LINK_DIRECTORIES(${LIBSDL2_LIBRARY_DIRS} ${LIBOPENTOK_LIBRARY_DIRS})
 
-ADD_EXECUTABLE(${PROJECT_NAME} main.cpp ${CMAKE_CURRENT_SOURCE_DIR}/../common/src/renderer.cpp)
-TARGET_LINK_LIBRARIES(${PROJECT_NAME} ${LIBSDL2_LIBRARIES} ${LIBOPENTOK_LIBRARIES})
+ADD_EXECUTABLE(${PROJECT_NAME} main.cpp ${CMAKE_CURRENT_SOURCE_DIR}/../common/src/otk_thread.c ${CMAKE_CURRENT_SOURCE_DIR}/../common/src/renderer.cpp)
+TARGET_LINK_LIBRARIES(${PROJECT_NAME} pthread ${LIBSDL2_LIBRARIES} ${LIBOPENTOK_LIBRARIES})
