+++ common/src/renderer.cpp
