+++ common/src/renderer.h
