--- wip/README.md
@@ -1,13 +1,0 @@
-# OpenTok Linux SDK samples
-
-These sample applications show how to use the OpenTok C API on Linux.
-
-## List of samples
-
-* The code in `connect-session` folder shows how to connect to an OpenTok session.
-* The code in `connect-session-with-settings` folder shows how to connect to an OpenTok session. It uses session settings.
-* The code in `session-signals` folder shows how to connect to an OpenTok session, send a signal and listen for new signals.
-* The code in `session-publish` folder shows how to connect to an OpenTok session and publish into a session.
-* The code in `session-publish-subscribe` folder shows how to connect to an OpenTok session, publish and subscribe.
-* The code in `audio-device` folder shows how to build and use a custom audio device.
-* The code in `video-capturer` folder shows how to build and use a custom video capturer.
